Refactor: #171 WeeklySummaryUiState 중첩 클래스 변경에 따른 프리뷰 참조 경로 수정
