+++ app/src/main/java/com/konkuk/medicarecall/ui/feature/statistics/weeklycard/WeeklyGlucoseCard.kt
@@ -23,7 +23,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.konkuk.medicarecall.ui.feature.statistics.viewmodel.WeeklyGlucoseUiState
+import com.konkuk.medicarecall.ui.feature.statistics.viewmodel.WeeklySummaryUiState
 import com.konkuk.medicarecall.ui.theme.LocalMediCareCallShadowProvider
 import com.konkuk.medicarecall.ui.theme.MediCareCallTheme
 import com.konkuk.medicarecall.ui.theme.figmaShadow
@@ -31,7 +31,7 @@ import com.konkuk.medicarecall.ui.theme.figmaShadow
 @Composable
 fun WeeklyGlucoseCard(
     modifier: Modifier = Modifier,
-    weeklyGlucose: WeeklyGlucoseUiState,
+    weeklyGlucose: WeeklySummaryUiState.WeeklyGlucoseUiState,
 ) {
     val hasBeforeMealData =
         weeklyGlucose.beforeMealNormal > 0 || weeklyGlucose.beforeMealHigh > 0 || weeklyGlucose.beforeMealLow > 0
@@ -191,7 +191,7 @@ private fun GlucoseStatusRow(
 @Composable
 fun PreviewWeeklyGlucoseCardRecorded() {
     WeeklyGlucoseCard(
-        weeklyGlucose = WeeklyGlucoseUiState(
+        weeklyGlucose = WeeklySummaryUiState.WeeklyGlucoseUiState(
             beforeMealNormal = 5,
             beforeMealHigh = 2,
             beforeMealLow = 1,
@@ -206,6 +206,6 @@ fun PreviewWeeklyGlucoseCardRecorded() {
 @Composable
 fun PreviewWeeklyGlucoseCardUnrecorded() {
     WeeklyGlucoseCard(
-        weeklyGlucose = WeeklyGlucoseUiState.EMPTY,
+        weeklyGlucose = WeeklySummaryUiState.WeeklyGlucoseUiState.EMPTY,
     )
 }
