+++ app/src/main/java/com/konkuk/medicarecall/ui/feature/statistics/weeklycard/WeeklyMentalCard.kt
@@ -23,7 +23,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.konkuk.medicarecall.R
-import com.konkuk.medicarecall.ui.feature.statistics.viewmodel.WeeklyMentalUiState
+import com.konkuk.medicarecall.ui.feature.statistics.viewmodel.WeeklySummaryUiState
 import com.konkuk.medicarecall.ui.theme.LocalMediCareCallShadowProvider
 import com.konkuk.medicarecall.ui.theme.MediCareCallTheme
 import com.konkuk.medicarecall.ui.theme.figmaShadow
@@ -31,7 +31,7 @@ import com.konkuk.medicarecall.ui.theme.figmaShadow
 @Composable
 fun WeeklyMentalCard(
     modifier: Modifier = Modifier,
-    mental: WeeklyMentalUiState,
+    mental: WeeklySummaryUiState.WeeklyMentalUiState,
 ) {
     Card(
         modifier = modifier
@@ -122,7 +122,7 @@ private fun MentalStatusRow(
 fun PreviewWeeklyMentalCardRecorded() {
     WeeklyMentalCard(
         modifier = Modifier.size(155.dp),
-        mental = WeeklyMentalUiState(
+        mental = WeeklySummaryUiState.WeeklyMentalUiState(
             good = 4,
             normal = 2,
             bad = 1,
@@ -135,6 +135,6 @@ fun PreviewWeeklyMentalCardRecorded() {
 fun PreviewWeeklyMentalCardUnrecorded() {
     WeeklyMentalCard(
         modifier = Modifier.size(155.dp),
-        mental = WeeklyMentalUiState.EMPTY,
+        mental = WeeklySummaryUiState.WeeklyMentalUiState.EMPTY,
     )
 }
